requirements.txt
Django==4.2
django-phone-field==1.8.1
djangorestframework==3.14.0
pytest==7.2.2
